feat(i3s): colorize by attribute
